Fibonacci Number

+++ Dynamic Programming/Fibonacci_Number.cpp
@@ -5,8 +5,7 @@
 
 using namespace std;
 
-
-long long fibb(long long n, unordered_map <long long, long long>& umap) {
+long long fibb(int n, unordered_map <int, long long>& umap) {
     if(umap.count(n) != 0)    return umap[n];
 
     else umap[n] = fibb(n - 1, umap) + fibb(n - 2, umap);
@@ -14,7 +13,7 @@ long long fibb(long long n, unordered_map <long long, long long>& umap) {
     return umap[n];
 }
 int main() {
-    unordered_map <long long, long long> umap;
+    unordered_map <int, long long> umap;
     umap[1] = 1; umap[2] = 1;
     int data;
     cin >> data;
